Organização do back e do front, sistema de envio de emails, correção de bug de inicialização, autenticação de usuário parcialmente feita

+++ BackendContratanti/src/main/java/com/projeto/BackendContratanti/Dto/EmpresaResponseDTO.java
@@ -2,9 +2,10 @@ package com.projeto.BackendContratanti.Dto;
 
 import com.projeto.BackendContratanti.Model.Empresa;
 
+
 import java.math.BigInteger;
 
-public record EmpresaResponseDTO (BigInteger id,String nome, String email, String telefone, String cnpj, String descricao, String url_linkedin, String senha){
+public record EmpresaResponseDTO (int id,String nome, String email, String telefone, String cnpj, String descricao, String url_linkedin, String senha){
     public EmpresaResponseDTO(Empresa empresa) {
         this(empresa.getId(), empresa.getNome(), empresa.getEmail(), empresa.getTelefone(), empresa.getUrl_linkedin(), empresa.getCnpj(), empresa.getSenha(), empresa.getDescricao());
     }
